Fix product ID parsing: parseInt → parseFloat (8 sites)
Product IDs in this app are floats (Date.now() + Math.random(), e.g.
1777019020364.0994). HTML checkbox/select values are strings, and
parseInt('1777019020364.0994') truncates the decimal to 1777019020364,
which never strictly-equals the original float. Lookups via includes()
or === silently return null/empty — manifesting as 'Error creating
benchmark', broken Compare flows, etc.

Changed 8 product-ID parseInt() → parseFloat() across:
- app.js (retest flow)
- consumer-panel-ui.js (2 sites)
- export-reporting-ui.js (3 sites)
- industry-benchmarks-ui.js (3 sites, incl. originally reported
  Create Custom Benchmark bug)

Slider value parseInt() calls (0–10 integer scale) NOT touched.

Also removes obsolete fix-benchmarks-create.js patch script.

SW VERSION: 3.4.13-fix-benchmarks-compare → 3.4.14-fix-product-id-parseint

+++ app.js
@@ -445,7 +445,7 @@ function handleFormSubmit(e) {
 
     // Check if this is a re-test
     const retestSelector = document.getElementById('retest-selector');
-    const originalTestId = retestSelector && retestSelector.value ? parseInt(retestSelector.value) : null;
+    const originalTestId = retestSelector && retestSelector.value ? parseFloat(retestSelector.value) : null;
     let testNumber = 1;
 
     if (originalTestId) {
